AWSCHAT-129 create live chat hook implementation
