+++ src/components/Chat/TollChat/TollChat.js
@@ -1,28 +1,14 @@
 'use client'
 
-import React, { useRef, useState, useEffect } from 'react'
+import React from 'react'
 
 import styles from './TollChat.module.scss'
-import { clearLocalStorage, getLocalStorage, isExpired, setLocalStorage } from '../../../lib/utils'
-import {
-  handleChatInit,
-  fetchAvailability,
-  startConversation,
-  endChat,
-  getConversationHistory,
-  postMessage,
-  listenToConversation
-} from '../../../../utils/chat/apis'
-import { convertTimeStamp, formatAgentImage, formatMessage, popNextUUID } from '../../../../utils/chat/libs'
 import ChatInput from './ChatInput'
 import { ChatForm } from '../ChatForm'
-import { validateChatForm } from '../utils/validateChatForm'
-
-import ChatMessageText from './ChatMessageText'
-import ChatMessageAttachment from './ChatMessageAttachment'
-import ChatMessageRichLink from './ChatMessageRichLink'
+import { LiveChatMessage } from './LiveChatMessage'
 import { HeaderButtons } from '../HeaderButtons'
 import { ConfirmationEndDialog } from '../ConfirmationEndDialog'
+import { useTollLiveChat } from '../hooks/useTollLiveChat'
 
 export const TollChat = ({
   availabilityAPI,
@@ -34,697 +20,65 @@ export const TollChat = ({
   chatStatus,
   chatRegion,
   setIsChatOpen = () => null,
-  isChatOpen, // this is to open chat from a button in the parent app instead of a floating head
+  isChatOpen,
   chatSms,
   trackChatEvent = () => null,
   chatClickedEventString = 'chatClicked',
   chatStartedEventString = 'chatStarted',
-  productCode, // ie JDE number of community/model/qmi
-  utils = {},
-  chatBotTransferData = null,
-  setChatBotTransferData = () => null,
-  setIsChatBotOpenExternal = () => null
+  productCode,
+  utils = {}
+  // chatBotTransferData = null,
+  // setChatBotTransferData = () => null,
+  // setIsChatBotOpenExternal = () => null
 }) => {
-  const isTransfering = useRef(false)
-  const isInConference = useRef(false)
-  const [showChatButton, setShowChatButton] = useState(false)
-  const [accessToken, setAccessToken] = useState(null)
-  const [messages, setMessages] = useState([])
-  const [conversationId, setConversationId] = useState(null)
-  const [showForm, setShowForm] = useState(false)
-  const [showChatHeader, setShowChatHeader] = useState(false)
-  const [showTextChatOptions, setShowTextChatOptions] = useState(false)
-  const [showWaitMessage, setShowWaitMessage] = useState(false)
-  const [showConfirmationEndMessage, setShowConfirmationEndMessage] = useState(false)
-  const chatContainerRef = useRef(null)
-  const [formData, setFormData] = useState({ name: '', email: '' })
-  const [isCurrentlyChatting, setIsCurrentlyChatting] = useState(false) // if therer is an active chat
-  const [showActiveTyping, setShowActiveTyping] = useState(false)
-  const [isMinimized, setIsMinimized] = useState(false)
-  const [systemMessage, setSystemMessage] = useState('') // system messages
-  const [chatPhoto, setChatPhoto] = useState(null) // osc image
-  const [agentName, setAgentName] = useState('Agent') // agent name
-  const [error, setError] = useState(null)
-  const [abortController, setAbortController] = useState(null)
-  const [isChatAvailabilityChecked, setIsChatAvailabilityChecked] = useState(null)
-  const [hasAgentEngaged, setHasAgentEngaged] = useState(false)
-  const [callbackUrl, setCallbackUrl] = useState(null)
-  const [unreadMessagesCount, setUnreadMessagesCount] = useState({
-    count: 0,
-    lastMessageId: null
+  const {
+    chatContainerRef,
+    showChatButton,
+    accessToken,
+    messages,
+    conversationId,
+    showForm,
+    showChatHeader,
+    showTextChatOptions,
+    showWaitMessage,
+    showConfirmationEndMessage,
+    formData,
+    isMinimized,
+    systemMessage,
+    chatPhoto,
+    error,
+    hasAgentEngaged,
+    callbackUrl,
+    unreadMessagesCount,
+    setFormData,
+    setError,
+    handleSubmit,
+    showTextChatOption,
+    showFormHandler,
+    handleMinimize,
+    handleConfirmationEnd,
+    handleStay,
+    handleEndChat
+  } = useTollLiveChat({
+    availabilityAPI,
+    endPoint,
+    apiSfOrgId,
+    apiSfName,
+    disableFloatingChatButton,
+    setChatStatus,
+    chatStatus,
+    chatRegion,
+    setIsChatOpen,
+    isChatOpen,
+    trackChatEvent,
+    chatClickedEventString,
+    chatStartedEventString,
+    productCode,
+    utils
+    // chatBotTransferData,
+    // setChatBotTransferData,
+    // setIsChatBotOpenExternal
   })
-
-  const initializeChat = async (firstName, lastName, email, isAgent, productCode, endPoint, apiSfOrgId, apiSfName) => {
-    try {
-      const token = await handleChatInit(endPoint, apiSfOrgId, apiSfName)
-
-      if (!token) throw new Error('No token received.')
-
-      setAccessToken(token.accessToken)
-
-      const newUuid = conversationId || crypto.randomUUID()
-      if (!conversationId || conversationId === null) {
-        setConversationId(newUuid)
-      }
-
-      const payload = {
-        accessToken: token.accessToken,
-        customerEmail: email,
-        customerFirstName: firstName || 'Guest',
-        customerLastName: lastName || '',
-        isAgent: isAgent,
-        conversationId: newUuid,
-        region: chatRegion,
-        endPoint,
-        apiSfOrgId,
-        apiSfName
-      }
-
-      if (productCode) {
-        payload.productCode = productCode
-      }
-
-      const response = await startConversation(payload)
-
-      if (response) {
-        let initialize = false
-        setConversationId(payload.conversationId)
-
-        const onSuccess = (abortController) => {
-          if (initialize) return
-
-          initialize = true
-          setLocalStorage(
-            'tbChat',
-            {
-              accessToken: token.accessToken,
-              conversationId: newUuid,
-              firstName,
-              lastName
-            },
-            1000 * 60 * 60 * 2 // expire in 2 hours
-          )
-
-          sendSystemtMessage({
-            accessToken: token.accessToken,
-            conversationId: newUuid,
-            message: '::System Message:: User initiated chat(' + location.href + ')'
-          })
-
-          setAbortController(abortController)
-        }
-
-        listenToConversation({
-          handleChatMessage,
-          onSuccess,
-          firstName,
-          lastName,
-          endPoint,
-          apiSfOrgId,
-          accessToken: token.accessToken,
-          conversationId: newUuid,
-          onError: () => {
-            console.log('Error starting conversation')
-            handleEndChat(token.accessToken, newUuid)
-            setShowForm(true)
-            setShowChatHeader(true)
-            setIsChatOpen(true)
-            setError('There was an error initializing the chat. Please try again.')
-          }
-        })
-      }
-    } catch (error) {
-      console.error('Error initializing chat')
-      handleEndChat()
-      setShowForm(true)
-      setShowChatHeader(true)
-      setIsChatOpen(true)
-      setError('There was an error initializing the chat. Please try again.')
-    }
-  }
-
-  const handleAddAgent = (entry) => {
-    setChatPhoto(formatAgentImage(entry.participant?.subject))
-    setSystemMessage(`You're chatting with ` + entry.displayName)
-    setAgentName(entry.displayName)
-    setShowWaitMessage(false)
-    setHasAgentEngaged(true)
-    setShowForm(false)
-    setIsCurrentlyChatting(true)
-  }
-
-  const handleChatMessage = async (event, firstName, lastName, accessToken, conversationId) => {
-    const messages = []
-    let message = {}
-    let data, messagePayload
-
-    switch (event.event) {
-      case 'ping':
-        break
-      case 'CONVERSATION_ROUTING_RESULT':
-        // seems to fire once when the the start conversation is called but agent has not
-        // accepted and no messages yet sent/delivered
-        data = JSON.parse(event.data)
-        messagePayload = JSON.parse(data.conversationEntry.entryPayload)
-
-        if (messagePayload?.routingType === 'Transfer') {
-          isTransfering.current = true
-        } else if (messagePayload?.routingType === 'Conference') {
-          isInConference.current = true
-        }
-
-        setIsCurrentlyChatting(true)
-        break
-      case 'CONVERSATION_PARTICIPANT_CHANGED':
-        data = JSON.parse(event.data)
-        messagePayload = JSON.parse(data.conversationEntry.entryPayload)
-        setShowWaitMessage(false)
-
-        if (
-          messagePayload?.entries?.[0]?.operation === 'remove' &&
-          messagePayload?.entries?.[0]?.participant?.role !== 'Supervisor'
-        ) {
-          // if (!isTransfering.current && !isInConference.current) {
-          // afterEndChatReset()
-          setSystemMessage(messagePayload.entries[0].displayName + ' left the conversation')
-          // } else {
-          // setSystemMessage(null)
-          // if (isTransfering.current) setShowWaitMessage(false)
-          // isTransfering.current = false
-          // isInConference.current = false
-          // }
-        } else {
-          for (let i = 0; i < messagePayload?.entries?.length; i++) {
-            const entry = messagePayload.entries[i]
-            if (entry.operation === 'add' && entry.participant?.role !== 'Supervisor') {
-              handleAddAgent(entry)
-              continue
-            }
-          }
-        }
-
-        break
-      case 'CONVERSATION_MESSAGE':
-        data = JSON.parse(event.data)
-
-        message = formatMessage(
-          {
-            ...data.conversationEntry,
-            entryPayload: JSON.parse(data.conversationEntry.entryPayload)
-          },
-          firstName,
-          lastName
-        )
-
-        if (
-          accessToken &&
-          conversationId &&
-          message.role === 'Agent' &&
-          message.payload?.formatType === 'Text' &&
-          message.payload?.text === '/url'
-        ) {
-          sendSystemtMessage({
-            accessToken: accessToken,
-            conversationId: conversationId,
-            message: '::System Message:: User on page: (' + location.href + ')'
-          })
-
-          return
-        }
-
-        messages.push(message)
-
-        break
-      case 'CONVERSATION_DELIVERY_ACKNOWLEDGEMENT':
-        console.log('conversation delivered acknowledge...')
-        break
-      case 'CONVERSATION_READ_ACKNOWLEDGEMENT':
-        console.log('conversation read acknowledge...')
-        break
-      case 'CONVERSATION_TYPING_STARTED_INDICATOR':
-        data = JSON.parse(event.data)
-        setShowActiveTyping(data)
-        break
-      case 'CONVERSATION_TYPING_STOPPED_INDICATOR':
-        setShowActiveTyping(false)
-        break
-      case 'CONVERSATION_CLOSE_CONVERSATION':
-        console.log('conversation close conversation...')
-        // afterEndChatReset()
-        break
-      default:
-        console.log('Unknown event:', event)
-        break
-    }
-
-    if (messages.length > 0) {
-      setMessages((prevMessages) => [...prevMessages, ...messages])
-    }
-  }
-
-  const handleSubmit = async (e) => {
-    const offlineMessage =
-      'Our apologies, but all members of our Online Sales Team are currently unavailable. Please try again later.'
-    e.preventDefault()
-    setError(null)
-    setSystemMessage(null)
-    setShowActiveTyping(false)
-
-    const form = e.target
-    const { valid, firstName, lastName } = validateChatForm(form)
-    if (!valid) return false
-
-    setShowWaitMessage(true)
-    setShowForm(false)
-    trackChatEvent(chatStartedEventString)
-
-    try {
-      const availability = await fetchAvailability(chatRegion, availabilityAPI)
-      if (availability?.data?.payload?.length > 0) {
-        const email = form.email?.value?.trim()
-        const isAgent = form.isAgent?.value ?? '0'
-
-        const { clientId, email_sha256, gaTrackId } = (await utils?.getGaTrackingIds?.(email)) || {}
-
-        setCallbackUrl(
-          `https://hello.tollbrothers.com/l/402642/2025-08-05/2chvs9x?email=${encodeURIComponent(
-            email
-          )}&fname=${encodeURIComponent(firstName)}&lname=${encodeURIComponent(
-            lastName
-          )}&gaClientId=${encodeURIComponent(clientId ?? '')}&gaUserId=${encodeURIComponent(
-            email_sha256 ?? ''
-          )}&gaTrackId=${encodeURIComponent(gaTrackId ?? '')}`
-        )
-
-        await initializeChat(firstName, lastName, email, isAgent, productCode, endPoint, apiSfOrgId, apiSfName)
-      } else {
-        setShowForm(true)
-        setShowWaitMessage(false)
-        setError(offlineMessage)
-      }
-    } catch (error) {
-      setShowForm(true)
-      setShowWaitMessage(false)
-      setError(offlineMessage)
-    }
-  }
-
-  const showTextChatOption = () => {
-    setShowTextChatOptions(!showTextChatOptions)
-  }
-
-  const showFormHandler = (trackEvent) => {
-    if (trackEvent) trackChatEvent(chatClickedEventString)
-    setIsChatOpen(true)
-    setShowChatHeader(true)
-    setShowForm(true)
-    setShowChatButton(false)
-  }
-
-  useEffect(() => {
-    if (!chatBotTransferData) {
-      if ((!chatRegion || chatStatus === 'offline') && !isCurrentlyChatting) {
-        setIsChatOpen(false)
-        setShowChatHeader(false)
-        setShowForm(false)
-        setShowChatButton(false)
-        setIsMinimized(false)
-        setShowConfirmationEndMessage(false)
-        setShowTextChatOptions(false)
-        return
-      }
-
-      if (isChatOpen && !isCurrentlyChatting) {
-        showFormHandler()
-        setShowConfirmationEndMessage(false)
-        setShowTextChatOptions(false)
-        // setIsMinimized(false)
-        return
-      }
-
-      if (
-        chatRegion &&
-        chatStatus === 'online' &&
-        !isCurrentlyChatting &&
-        !disableFloatingChatButton &&
-        !showChatHeader
-      ) {
-        setShowChatButton(true)
-      } else {
-        setShowChatButton(false)
-      }
-    }
-  }, [chatStatus, isCurrentlyChatting, disableFloatingChatButton, isChatOpen, chatRegion, chatBotTransferData])
-
-  useEffect(() => {
-    async function getOscInfo() {
-      try {
-        const availability = await fetchAvailability(chatRegion, availabilityAPI)
-        if (availability?.data?.payload?.length > 0) {
-          setChatStatus('online')
-          const index = Math.floor(Math.random() * availability.data.payload.length)
-          if (!isCurrentlyChatting) {
-            setChatPhoto(availability.data.payload[index]?.photo)
-          }
-        }
-        setIsChatAvailabilityChecked(true)
-      } catch (error) {
-        setIsChatAvailabilityChecked(true)
-        console.error('Error fetching osc data:', error)
-      }
-    }
-
-    if (chatRegion) {
-      getOscInfo()
-    } else {
-      // if no chatRegion (pages with no OSC) just set this to true.
-      setIsChatAvailabilityChecked(true)
-    }
-
-    setChatStatus('offline')
-  }, [chatRegion, availabilityAPI])
-
-  const sendSystemtMessage = async ({ accessToken, conversationId, message }) => {
-    try {
-      await postMessage({
-        accessToken,
-        conversationId,
-        msg: message,
-        endPoint,
-        apiSfName
-      })
-    } catch {
-      console.error('Error sending system message.')
-    }
-  }
-
-  const getConversationList = async (tbChat) => {
-    try {
-      const response = await getConversationHistory({
-        accessToken: tbChat.accessToken,
-        conversationId: tbChat.conversationId,
-        endPoint
-      })
-
-      setShowWaitMessage(true)
-      if (response?.conversationEntries?.length > 0) {
-        // return only messages
-        const messages = response.conversationEntries.filter((entry) => {
-          return entry.entryType === 'Message'
-        })
-        const formattedMessages = messages.map((message) => {
-          return formatMessage(message, tbChat.firstName, tbChat.lastName)
-        })
-
-        let chatWasEndedByAgentWhileOffline = false
-        const tbChatBackup = { ...tbChat }
-
-        const hasTransferEvent = response.conversationEntries.some((entry) => {
-          return entry.entryType === 'RoutingResult' && entry.entryPayload?.routingType === 'Transfer'
-        })
-
-        const hasConferenceEvent = response.conversationEntries.some((entry) => {
-          return entry.entryType === 'RoutingResult' && entry.entryPayload?.routingType === 'Conference'
-        })
-
-        response.conversationEntries.map((entry) => {
-          if (entry.entryType === 'ParticipantChanged' && entry.entryPayload?.entries?.length > 0) {
-            entry.entryPayload.entries.map((entry) => {
-              // find the add entry to get agent name
-              if (entry.operation === 'add' && entry.participant?.role !== 'Supervisor') {
-                // always first since the messages are in chronological order.
-                // If the user was transfered this will be in the message list a second time
-                // so we need to add all the chat data back into localStorage and React state
-                handleAddAgent(entry)
-                setLocalStorage(
-                  // restoring chat data in case there was a transfer
-                  'tbChat',
-                  {
-                    accessToken: tbChatBackup.accessToken,
-                    conversationId: tbChatBackup.conversationId,
-                    firstName: tbChatBackup.firstName,
-                    lastName: tbChatBackup.lastName
-                  },
-                  1000 * 60 * 60 * 2 // expire in 2 hours
-                )
-                setAccessToken(tbChatBackup.accessToken)
-                setConversationId(tbChatBackup.conversationId)
-                chatWasEndedByAgentWhileOffline = false
-              } else if (
-                entry.operation === 'remove' &&
-                entry.participant?.role !== 'Supervisor' &&
-                !hasTransferEvent &&
-                !hasConferenceEvent
-              ) {
-                // always after the add event since the messages are in chronological order
-                // so if there was no transfer this will be the last event and we can display the agent left messsage
-
-                // afterEndChatReset()
-                setSystemMessage(entry.displayName + ' left the conversation')
-                setHasAgentEngaged(false)
-                chatWasEndedByAgentWhileOffline = true
-              }
-            })
-          }
-        })
-
-        if (!chatWasEndedByAgentWhileOffline) {
-          setMessages(formattedMessages ?? [])
-        }
-      } else {
-        throw new Error()
-      }
-    } catch (err) {
-      console.error('Error retreiving chat history')
-    }
-  }
-
-  const reestablishConnection = (event) => {
-    if (event && event.type === 'visibilitychange' && document.hidden) {
-      return
-    }
-
-    let initialize = false
-    const isTabVisiblilityEvent = event && event.type === 'visibilitychange'
-    const tbChat = getLocalStorage('tbChat') ?? chatBotTransferData
-
-    if (!tbChat || isExpired(tbChat.expiry)) {
-      clearLocalStorage('tbChat')
-      setIsChatBotOpenExternal(false)
-      setChatBotTransferData(null)
-      return
-    }
-
-    const value = tbChat.value ?? tbChat
-    if (!value?.conversationId || !value?.accessToken) {
-      return
-    }
-
-    listenToConversation({
-      handleChatMessage,
-      firstName: value.firstName,
-      lastName: value.lastName,
-      endPoint,
-      apiSfOrgId,
-      accessToken: value.accessToken,
-      conversationId: value.conversationId,
-      onError: () => {
-        handleEndChat(value.accessToken, value.conversationId)
-      },
-      onSuccess: (abortController) => {
-        if (initialize) return
-
-        initialize = true
-        setAccessToken(value.accessToken)
-        setConversationId(value.conversationId)
-        setShowChatButton(false)
-        setIsCurrentlyChatting(true)
-        setSystemMessage(null)
-        setShowActiveTyping(false)
-        setShowChatHeader(true)
-        setShowTextChatOptions(false)
-        getConversationList(value)
-
-        if (chatBotTransferData) {
-          setLocalStorage(
-            // storing chat data since there was a transfer from chatbot
-            'tbChat',
-            {
-              accessToken: value.accessToken,
-              conversationId: value.conversationId,
-              firstName: value.firstName,
-              lastName: value.lastName
-            },
-            1000 * 60 * 60 * 2 // expire in 2 hours
-          )
-        } else {
-          setShowWaitMessage(false)
-        }
-
-        // turned this off since it was sending a system message every time the tab was restored
-        // can confuse agents
-        // if (!isTabVisiblilityEvent) {
-        //   sendSystemtMessage({
-        //     accessToken: value.accessToken,
-        //     conversationId: value.conversationId,
-        //     message: '::System Message:: Guest restored connection'
-        //   })
-        // }
-        setAbortController(abortController)
-        setIsChatOpen(true)
-      }
-    })
-  }
-
-  useEffect(() => {
-    window.addEventListener('visibilitychange', reestablishConnection)
-
-    return () => {
-      window.removeEventListener('visibilitychange', reestablishConnection)
-    }
-  }, [])
-
-  useEffect(() => {
-    // need to wait for availabilty api to return before reestablishing connection
-    // else we might show chat input too soon (race condition)
-    if (isChatAvailabilityChecked) {
-      reestablishConnection()
-    }
-  }, [isChatAvailabilityChecked])
-
-  useEffect(() => {
-    if (chatBotTransferData) {
-      setShowChatButton(false)
-      setShowChatHeader(true)
-      setShowWaitMessage(true)
-      setShowTextChatOptions(false)
-      setIsChatOpen(true)
-    }
-  }, [chatBotTransferData])
-
-  const handleMinimize = () => {
-    setError(null)
-    setIsMinimized(!isMinimized)
-    setShowConfirmationEndMessage(false)
-    setUnreadMessagesCount({ count: 0, lastMessageId: null })
-  }
-
-  const handleConfirmationEnd = () => {
-    setError(null)
-    setShowConfirmationEndMessage(true)
-  }
-
-  const handleStay = () => {
-    setError(null)
-    setShowConfirmationEndMessage(false)
-  }
-
-  const afterEndChatReset = () => {
-    setError(null)
-    clearLocalStorage('tbChat')
-    setIsMinimized(false)
-    setShowChatButton(false)
-    setIsCurrentlyChatting(false)
-    setHasAgentEngaged(false)
-    setConversationId(null)
-    // setSystemMessage(null)
-    setShowActiveTyping(false)
-    setShowForm(false)
-    setAccessToken(null)
-    setMessages([])
-    setShowWaitMessage(false)
-    setShowConfirmationEndMessage(false)
-    setShowTextChatOptions(false)
-    setAgentName('Agent')
-    setSystemMessage(null)
-    setUnreadMessagesCount({ count: 0, lastMessageId: null })
-    setChatBotTransferData(null)
-    setIsChatBotOpenExternal(false)
-    isTransfering.current = false
-    isInConference.current = false
-    if (abortController) {
-      // closes sse connection
-      abortController.abort()
-      setAbortController(null)
-    }
-  }
-
-  const handleEndChat = async (accessToken, conversationId) => {
-    afterEndChatReset()
-    setShowChatHeader(false)
-    setIsChatOpen(false)
-
-    if (!accessToken || !conversationId) {
-      return
-    }
-
-    try {
-      console.log('Ending chat...')
-      await endChat({
-        accessToken,
-        conversationId,
-        endPoint,
-        apiSfName
-      })
-    } catch (error) {
-      console.error('Chat end error:', error.message)
-    }
-  }
-
-  useEffect(() => {
-    if (chatContainerRef.current) {
-      chatContainerRef.current.scrollTop = chatContainerRef.current.scrollHeight
-    }
-  }, [messages, isMinimized])
-
-  useEffect(() => {
-    const typingMessageExists = messages.some((message) => {
-      return message.payload?.formatType === 'Typing'
-    })
-
-    if (typingMessageExists && showActiveTyping) return
-
-    if (showActiveTyping) {
-      // add typing message to message list
-      setMessages((prevMessages) => [
-        ...prevMessages,
-        {
-          id: popNextUUID(),
-          type: 'Message',
-          text: showActiveTyping?.conversationEntry?.senderDisplayName + ' is typing...',
-          payload: { formatType: 'Typing' },
-          role: 'Agent'
-        }
-      ])
-    } else {
-      setMessages((prevMessages) => {
-        // remove typing message from message list
-        return prevMessages.filter((message) => {
-          return message.payload?.formatType !== 'Typing'
-        })
-      })
-    }
-  }, [showActiveTyping])
-
-  useEffect(() => {
-    const lastMessage = messages[messages.length - 1]
-
-    if (
-      !document.hidden &&
-      lastMessage?.id !== unreadMessagesCount.lastMessageId &&
-      lastMessage?.payload?.formatType !== 'Typing' &&
-      lastMessage?.type === 'Message' &&
-      lastMessage?.role === 'Agent' &&
-      isMinimized
-    ) {
-      setUnreadMessagesCount((prev) => {
-        return {
-          count: prev.count + 1,
-          lastMessageId: lastMessage?.id
-        }
-      })
-    }
-  }, [messages])
 
   return (
     <div
@@ -742,7 +96,7 @@ export const TollChat = ({
               <div className={styles.textChatWrapper}>
                 <button
                   className={`${styles.chatButton} ${styles.textChatButtons}`}
-                  onClick={!isMinimized ? showFormHandler : handleMinimize}
+                  onClick={!isMinimized ? () => showFormHandler(true) : handleMinimize}
                 >
                   <img src='https://cdn.tollbrothers.com/sites/comtollbrotherswww/svg/chat.svg' alt='chat' />
                   Chat
@@ -827,38 +181,16 @@ export const TollChat = ({
       <div className={styles.messagesWrapper} ref={chatContainerRef}>
         {!isMinimized && (
           <>
-            {messages.map(
-              (message, index) =>
-                message.type === 'Message' &&
-                !message.text?.includes('::System Message::') &&
-                !(message.text?.startsWith('/url') && message?.role === 'Agent') && (
-                  <React.Fragment key={message.id}>
-                    <div className={styles.timestamp}>{convertTimeStamp(message.timestamp)}</div>
-                    {message.payload?.formatType === 'RichLink' && (
-                      <>
-                        {message.text && <ChatMessageText message={message} />}
-                        <ChatMessageRichLink
-                          richLink={message.payload}
-                          leftAlign={message?.role === 'Agent' || message?.role === 'System'}
-                        />
-                      </>
-                    )}
-                    {message.payload?.formatType === 'Attachments' && (
-                      <>
-                        {message.text && <ChatMessageText message={message} />}
-                        <ChatMessageAttachment
-                          attachments={message?.payload?.attachments}
-                          hasText={Boolean(message.text)}
-                          leftAlign={message.role === 'Agent'}
-                        />
-                      </>
-                    )}
-                    {(message.payload?.formatType === 'Text' || message.payload?.formatType === 'Typing') && (
-                      <ChatMessageText message={message} />
-                    )}
-                  </React.Fragment>
-                )
-            )}
+            {messages
+              .filter(
+                (message) =>
+                  message.type === 'Message' &&
+                  !message.text?.includes('::System Message::') &&
+                  !(message.text?.startsWith('/url') && message?.role === 'Agent')
+              )
+              .map((message) => (
+                <LiveChatMessage key={message.id} message={message} />
+              ))}
           </>
         )}
       </div>
