+++ src/components/Chat/TollChat/TollChat.module.scss
@@ -109,14 +109,6 @@
     flex-shrink: 0;
 }
 
-.timestamp {
-    font-size: 10px;
-    font-family: $special_font_family;
-    white-space: nowrap;
-    color: #000;
-    text-align: center;
-}
-
 .textChatOptions {
     position: absolute;
     bottom: 0;
